Added vostok gateway send feature

+++ app/src/main/java/com/wavesplatform/wallet/v2/data/model/remote/response/gateway/InitDepositResponse.kt
@@ -12,9 +12,9 @@ data class InitDepositResponse(
     @SerializedName("address")
     var address: String,
     @SerializedName("fee")
-    var fee: Int,
+    var fee: Long,
     @SerializedName("maxAmount")
-    var maxAmount: Int,
+    var maxAmount: Long,
     @SerializedName("minAmount")
-    var minAmount: Int
+    var minAmount: Long
 )
